assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    # 4110 only:
     "cython>=3",
]
build-backend = "setuptools.build_meta"

[project]
version = "1.0.0"
requires-python = ">=3.8"
license = {text = "MIT License"}
name = "instapy"
authors = [{"name" = "Eyyüb Güven", "email" = "eyyubg@uio.no"}] 
description = "A package for image processing and analysis"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
 ]


[project.scripts]
instapy = "in3110_instapy.cli:main"